+++ Samples/invaders/src/DecoratorStarfield.cpp
@@ -123,8 +123,8 @@ void DecoratorStarfield::RenderElement(Rocket::Core::Element* ROCKET_UNUSED_PARA
 
 void DecoratorStarfield::StarField::Update()
 {
-	float time = Shell::GetElapsedTime();
-	float delta_time = time - last_update;
+	double time = Shell::GetElapsedTime();
+	float delta_time = float(time - last_update);
 	last_update = time;
 
 	if (!GameDetails::GetPaused())
