+++ Samples/shell/src/ShellSystemInterface.cpp
@@ -29,7 +29,7 @@
 #include <Shell.h>
 
 // Get the number of seconds elapsed since the start of the application
-float ShellSystemInterface::GetElapsedTime()
+double ShellSystemInterface::GetElapsedTime()
 {
 	return Shell::GetElapsedTime();
 }
